+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## Changes in 1.9.3
 
-- Small performance improvement in `longForm()` implementation.
+- Small performance improvement in `longForm()` implementation: avoid using
+  `findMatches()` to order the result if not needed (i.e., if duplicated spectra
+  are requested).
 
 ## Changes in 1.9.2